Renaming server module to transport
